requirements.txt
matplotlib==3.7.1
numpy==1.24.4
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.68
Pillow==10.3.0
plyfile==1.0.3
tensorboardX==2.6
#tensorboardX==2.6.2.2
torch==1.12.0
torchvision==0.13.0
